chore: move to discord/commands and discord/interactions
