+++ src/discord/commands/pdf.ts
@@ -1,9 +1,9 @@
 import { SlashCommandBuilder } from 'discord.js'
-import { pdf2Markdown } from '../queues'
+import { downloadPDF } from '../../queues'
 
 export default {
   data: new SlashCommandBuilder()
-    .setName('parse')
+    .setName('pdf')
     .addStringOption((option) =>
       option.setName('url').setDescription('URL to PDF file').setRequired(true)
     )
@@ -16,7 +16,7 @@ export default {
     const url = interaction.options.getString('url')
     if (!url) {
       await interaction.followUp({
-        content: 'No url provided. Try again with /parse <url>',
+        content: 'No url provided. Try again with /pdf <url>',
         ephemeral: true,
       })
 
@@ -24,14 +24,13 @@ export default {
     }
 
     const reply = await interaction.reply({
-      content:
-        'Tack! Nu är din årsredovisning placerad i kö för hantering av LLama.',
+      content: 'Tack! Nu är din årsredovisning placerad i kö.',
       fetchReply: true,
     })
     const channelId = interaction.channelId
     const messageId = reply.id
 
-    pdf2Markdown.add('parse pdf ' + url.slice(-20), {
+    downloadPDF.add('download pdf ' + url.slice(-20), {
       url,
       channelId,
       messageId,
