+++ src/discord/commands/ping.ts
