--- src/interactions/reject.ts
@@ -1,10 +1,0 @@
-export default {
-  async execute(interaction, job) {
-    const { url, channelId, messageId, existingId, existingPdfHash } = job.data
-    interaction.update({
-      content: 'Rejected!',
-      embeds: [],
-      components: [],
-    })
-  },
-}
